+++ docs/src/app/components/pages/components/Stepper/CustomStepConnector.js
@@ -29,26 +29,26 @@ class CustomStepConnector extends React.Component {
       case 0:
         return (
           <p>
-            For each ad campaign that you create, you can control how much you're willing to
-            spend on clicks and conversions, which networks and geographical locations you want
-            your ads to show on, and more.
+            {'For each ad campaign that you create, you can control how much you\'re willing to ' +
+            'spend on clicks and conversions, which networks and geographical locations you want ' +
+            'your ads to show on, and more.'}
           </p>
         );
 
       case 1:
         return (
           <p>
-            An ad group contains one or more ads which target a shared set of keywords.
+            {'An ad group contains one or more ads which target a shared set of keywords.'}
           </p>
         );
 
       case 2:
         return (
           <p>
-            Try out different ad text to see what brings in the most customers, and learn how
-            to enhance your ads using features like ad extensions. If you run into any problems
-            with your ads, find out how to tell if they're running and how to resolve approval
-            issues.
+            {'Try out different ad text to see what brings in the most customers, and learn ' +
+            'how to enhance your ads using features like ad extensions. If you run into any ' +
+            'problems with your ads, find out how to tell if they\'re running and how to ' +
+            'resolve approval issues.'}
           </p>
         );
     }
